pyproject.toml
[project]
name = "vsm"
version = "0.1.0"
description = "VSM core with ports and adapters"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["vsm*"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
